+++ src/index.js
@@ -1,2 +1,3 @@
 export { auditRepository, formatMarkdownReport } from "./audit.js";
+export { CONFIG_FILE, loadConfig } from "./config.js";
 export { rules } from "./rules.js";